requirements.txt
asgiref==3.7.2
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
defusedxml==0.8.0rc2
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-jwt==1.19.2
ecdsa==0.18.0
idna==3.6
oauthlib==3.2.2
pip-autoremove==0.10.0
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
